+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased (0.15.2)
+# 0.15.2
 * Add `FaceMut::set_variation` trait abstraction for calling mutable `Face::set_variation` via `OwnedFace`.
 * Use edition 2021.